Moved unpacking map packs from application to activity
Prior to this fix, the spinner spun forever in the test in the testing section below.  In order to make the venue appear, it required restarting the app.

The root cause was a timing problem.  The venue tried to render before the map pack finished unpacking.

The solution was to call getVenueList() and loadVenue() when the map pack finished loading.  In order for that to happen, unpacking had to be done in the MainActivity instead of MyApplication.

Also updated dependencies.

Testing:
Given the app has been deleted from the device
  And the device is set to airplane mode
  And the command "rm app/src/main/assets/locuslabs/android-*.tar.xz" is run in the root directory of the app
  And the spinner spins forever when user launches app from Android Studio
  And a new map pack was copied to app/src/main/assets/locuslabs
When the user launches app from Android Studio
Then the a venue should appear